Feature/userdefaults scoping (#27)
* Make use of `UserDefaults` with a specific suitname

Also prefixed some keys with the plugin identifier. Addresses #24

* Cleanup

* Make optionality more explicit

`FlutterCallbackCache.lookupCallbackInformation` may return nil in certain cases. Force unwrapping the `flutterCallbackInformation` makes this more explicit.

Maybe a graceful exit if `flutterCallbackInformation` is nil would be better, but to be discussed and adressed in a different PR.

* Don't expose user defaults key

Also no need for staticness